created separate component for drozpone
